requirements.txt
python-pptx
deep-translator